+++ cli/src/cli.rs
@@ -9,7 +9,7 @@ use crate::utils::docker::{CommandType, Container};
 use crate::command::build::compose_build;
 use crate::command::cd::{cd_project, exec_cd_project};
 use crate::command::completion::{exec_shell_completion, shell_completion};
-use crate::command::config::{check_config, exec_check_config};
+use crate::command::config::{check_config, exec_check_config, register_config, exec_register_config, unregister_config, exec_unregister_config};
 use crate::command::create::compose_create;
 use crate::command::down::compose_down;
 use crate::command::events::compose_events;
@@ -68,6 +68,8 @@ fn cli() -> Command {
         .subcommand(shell_completion())
         .subcommand(cd_project())
         .subcommand(check_config())
+        .subcommand(register_config())
+        .subcommand(unregister_config())
         .subcommand(projects_infos())
 }
 
@@ -80,6 +82,8 @@ pub async fn run(container: &dyn Container, config: &mut dyn CliConfig) -> Resul
     match command_name {
         "infos" => exec_projects_infos(config, container).await?,
         "check-config" => exec_check_config(config)?,
+        "register" => exec_register_config(config)?,
+        "unregister" => exec_unregister_config(config)?,
         "completion" => exec_shell_completion(&mut cli(), args)?,
         _ => {}
     }